adding i18n from rebase to correct files
