+++ src/locales/en/translation.json
@@ -3,7 +3,7 @@
     "amount": {
       "header": "100% of your support will go to the merchant toward their operating costs. Your transaction is secure, encrypted and directly goes to the owner for sustaining their business.",
       "light_up_chinatown": "100% of donation goes towards permanent light fixtures throughout Chinatown.",
-      "mega_gam":"Please add your payment information below.",
+      "mega_gam": "Please add your payment information below.",
       "label1": "Select amount",
       "label2": "Other:",
       "minimum": "Minimum",
@@ -151,7 +151,7 @@
         "luc_address": "Shipping Address: "
       }
     },
-    "modalConfirmation" : {
+    "modalConfirmation": {
       "mega_gam_line_1": "You’ve lent a helping hand to community members and small businesses when they needed it most. Together, we can back those most impacted by the pandemic and keep the city’s Chinatowns vibrant.",
       "mega_gam_line_2": "You will receive an email receipt shortly with the details of your donation."
     }
